Fix schedule notification bug
